requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
WTForms==2.3.1
Werkzeug==1.0.1
bcrypt==3.1.7
biopython==1.76
bitarray==1.2.1
mmh3==2.5.1
numpy==1.18.2
pandas==1.0.3
requests==2.23.0
scikit-learn==0.23.1